ultimo cambio

+++ src/main/java/org/iesvdm/transformer/JoinByAdding.java
@@ -5,6 +5,6 @@ class JoinByAdding implements Joiner<Integer>
 
     public Integer join(Integer int1,Integer int2)
     {
-        return int1+int2;
+        return (Integer) (int1+int2);
     }
 }
